requirements.txt
flask
flask_htmx
flask_bcrypt
flask_socketio
pymongo
PyYAML
kubernetes==28.1.0
pyzmq
requests
